# Conflicts:
#   backend/timing/src/main/java/com/kkukku/timing/TimingApplication.java
#   backend/timing/src/main/java/com/kkukku/timing/security/configs/SecurityConfig.java
#   backend/timing/src/main/resources/application.yml

+++ backend/timing/src/main/java/com/kkukku/timing/apis/member/services/MemberService.java
@@ -30,25 +30,17 @@ public class MemberService {
     public void updateMember(Integer memberId, MemberUpdateRequest memberUpdateRequest,
         MultipartFile multipartFile) {
 
+        MemberEntity member = getMemberById(memberId);
         if (memberUpdateRequest != null) {
-            memberRepository.findById(memberId)
-                            .ifPresent(
-                                selectMember -> {
-                                    selectMember.setNickname(memberUpdateRequest.getNickname());
-                                    memberRepository.save(selectMember);
-                                });
+            member.setNickname(memberUpdateRequest.getNickname());
+            memberRepository.save(member);
         }
 
         if (multipartFile != null) {
             String profileImageUrl = "/" + s3Service.uploadFile(multipartFile);
-            memberRepository.findById(memberId)
-                            .ifPresent(
-                                selectMember -> {
-                                    selectMember.setProfileImageUrl(profileImageUrl);
-                                    memberRepository.save(selectMember);
-                                });
+            member.setProfileImageUrl(profileImageUrl);
+            memberRepository.save(member);
         }
-
     }
 
     public MemberDetailResponse getMemberInfo(String memberEmail) {
@@ -64,7 +56,9 @@ public class MemberService {
         MemberEntity member = getMemberById(memberId);
 
         String profileImgUrl = member.getProfileImageUrl();
-        s3Service.deleteFile(profileImgUrl);
+        if (!profileImgUrl.startsWith("http")) {
+            s3Service.deleteFile(profileImgUrl);
+        }
 
         member.delete();
